P3: make configuration parser ctype domains byte-safe
This repairs a current C/POSIX argument-domain defect with a potential runtime trigger. No production failure is known, but high-bit memory bytes promote negatively on signed-plain-char targets and are neither EOF by provenance nor representable as unsigned char.

Convert exactly the ten assigned isalnum, isalpha, isdigit, and isspace arguments to Netdata's established uint8_t boundary convention. This preserves each original byte expression and removes undefined behavior without adding a parser abstraction.

Complete caller tracing covers systemd Cgroup detection, both privileged Cgroup path-validation flows, eBPF hostname/IP/port INI and Function parsing plus tests, WebRTC ICE startup configuration, and Dynamic Configuration function registration. Token termination, pointer movement, enum and flag selection, list mutation, cap enforcement, locking, allocation, ownership, cleanup, and returns remain unchanged.

Valid ASCII grammar and all public behavior are unchanged. High-bit bytes are now classified deterministically under the existing C locale. The privileged path allow-list is not widened, and the change adds no pass, branch, allocation, I/O, lock, or meaningful runtime cost.

Validated at low priority with GCC and Clang in default, signed, unsigned, and UBSan modes; compiled all five affected translation units including the active WebRTC branch; and passed exact source and diff checks. The final structured six-reviewer gate passed independent raw-process and semantic audits with exact production-grade verdicts.

(cherry picked from commit 296542d6cfea4fa9834e90bb9892806a1b10dfd5)

+++ src/collectors/cgroups.plugin/cgroup-network.c
@@ -704,7 +704,7 @@ int verify_path(const char *path) {
 
     const char *s = path;
     while(*s != '\0') {
-        if (isalnum(*s) || is_valid_path_symbol(*s))
+        if (isalnum((uint8_t)*s) || is_valid_path_symbol(*s))
             s += 1;
         else if (*s == '\\' && is_valid_hex_escape(s))
             s += 4;
